+++ engine/flink/management/sample/src/main/scala/pl/touk/nussknacker/engine/management/sample/global/ConfigDynamicGlobalVariable.scala
@@ -1,7 +1,7 @@
 package pl.touk.nussknacker.engine.management.sample.global
 
 import pl.touk.nussknacker.engine.api.MetaData
-import pl.touk.nussknacker.engine.api.typed.DynamicGlobalVariable
+import pl.touk.nussknacker.engine.api.typed.{DynamicGlobalVariable, TypedMap}
 import pl.touk.nussknacker.engine.api.typed.typing.{Typed, TypingResult}
 
 /**
@@ -11,13 +11,12 @@ object ConfigDynamicGlobalVariable extends DynamicGlobalVariable {
 
   import scala.collection.JavaConverters._
 
-  private val configurations = {
-    Map(
-      "prod" -> List(Map("a" -> 1, "b" -> "B").asJava, Map("a" -> 2, "b" -> "BB").asJava).asJava
+  private val configurations = Map(
+      "prod" -> List(TypedMap(Map("a" -> 1, "b" -> "B")), TypedMap(Map("a" -> 2, "b" -> "BB"))).asJava
     ).withDefaultValue(
-      List(Map[String, Any]("a" -> 1).asJava, Map[String, Any]("a" -> 2).asJava, Map[String, Any]("a" -> 3).asJava).asJava
+      List(TypedMap(Map[String, Any]("a" -> 1)), TypedMap(Map[String, Any]("a" -> 2)), TypedMap(Map[String, Any]("a" -> 3))).asJava
     )
-  }
+
   override def value(metadata: MetaData): Any = {
     configurations(readProperty(metadata))
   }